requirements.txt
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.13.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachelib==0.13.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
Deprecated==1.3.1
distro==1.9.0
docstring_parser==0.18.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
Flask-Limiter==3.9.0
Flask-Session==0.8.0
Flask-WTF==1.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.12
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
limits==5.8.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgspec==0.21.1
ordered-set==4.1.0
packaging==26.1
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
python-dotenv==1.2.2
requests==2.33.1
rich==13.9.4
sniffio==1.3.1
soupsieve==2.8.3
stripe==11.1.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
wrapt==2.1.2
WTForms==3.2.1
